web/requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
psycopg2==2.9.3  # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django-anymail==8.6  # https://github.com/anymail/django-anymail
# minify html/css/js
django-minify-html==1.3.0
# sentry error tracking for production
sentry_sdk==1.9.8